refactor: allPokemons, seenPokemons 통합
