--- app/src/main/java/com/sdhong/pokemonapp/model/AllPokemons.kt
@@ -1,84 +1,0 @@
-package com.sdhong.pokemonapp.model
-
-val allPokemons = listOf(
-    Pokemon.Normal(
-        name = "bulbasaur",
-        imgUrl = "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/1.png"
-    ),
-    Pokemon.Normal(
-        name = "ivysaur",
-        imgUrl = "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/2.png"
-    ),
-    Pokemon.Normal(
-        name = "venusaur",
-        imgUrl = "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/3.png"
-    ),
-    Pokemon.Normal(
-        name = "charmander",
-        imgUrl = "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/4.png"
-    ),
-    Pokemon.Normal(
-        name = "charmeleon",
-        imgUrl = "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/5.png"
-    ),
-    Pokemon.Normal(
-        name = "charizard",
-        imgUrl = "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/6.png"
-    ),
-    Pokemon.Normal(
-        name = "squirtle",
-        imgUrl = "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/7.png"
-    ),
-    Pokemon.Normal(
-        name = "wartortle",
-        imgUrl = "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/8.png"
-    ),
-    Pokemon.Normal(
-        name = "blastoise",
-        imgUrl = "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/9.png"
-    ),
-    Pokemon.Normal(
-        name = "caterpie",
-        imgUrl = "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/10.png"
-    ),
-    Pokemon.Normal(
-        name = "metapod",
-        imgUrl = "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/11.png"
-    ),
-    Pokemon.Normal(
-        name = "butterfree",
-        imgUrl = "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/12.png"
-    ),
-    Pokemon.Normal(
-        name = "weedle",
-        imgUrl = "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/13.png"
-    ),
-    Pokemon.Normal(
-        name = "kakuna",
-        imgUrl = "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/14.png"
-    ),
-    Pokemon.Normal(
-        name = "beedrill",
-        imgUrl = "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/15.png"
-    ),
-    Pokemon.Normal(
-        name = "pidgey",
-        imgUrl = "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/16.png"
-    ),
-    Pokemon.Normal(
-        name = "pidgeotto",
-        imgUrl = "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/17.png"
-    ),
-    Pokemon.Normal(
-        name = "pidgeot",
-        imgUrl = "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/18.png"
-    ),
-    Pokemon.Normal(
-        name = "rattata",
-        imgUrl = "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/19.png"
-    ),
-    Pokemon.Normal(
-        name = "raticate",
-        imgUrl = "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/20.png"
-    )
-)
