--- app/src/main/java/com/sdhong/pokemonapp/model/SeenPokemons.kt
@@ -1,3 +1,0 @@
-package com.sdhong.pokemonapp.model
-
-val seenPokemons = mutableListOf<Pokemon.Seen>()
